# cardinal/server/docs/swagger.json
{
    "schemes": [
        "http",
        "ws"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend server for World Engine",
        "title": "Cardinal",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/debug/state": {
            "post": {
                "description": "Retrieves a list of all entities in the game state",
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves a list of all entities in the game state",
                "responses": {
                    "200": {
                        "description": "List of all entities",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg_world_dev_world-engine_cardinal_types.EntityData"
                            }
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "Establishes a new websocket connection to retrieve system events",
                "produces": [
                    "application/json"
                ],
                "summary": "Establishes a new websocket connection to retrieve system events",
                "responses": {
                    "101": {
                        "description": "Switch protocol to ws",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Retrieves the status of the server and game loop",
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves the status of the server and game loop",
                "responses": {
                    "200": {
                        "description": "Server and game loop status",
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.GetHealthResponse"
                        }
                    }
                }
            }
        },
        "/query/receipts/list": {
            "post": {
                "description": "Retrieves all transaction receipts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves all transaction receipts",
                "parameters": [
                    {
                        "description": "Query body",
                        "name": "GetReceiptsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.GetReceiptsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of receipts",
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.GetReceiptsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/query/{queryGroup}/{queryName}": {
            "post": {
                "description": "Executes a query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Executes a query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query group",
                        "name": "queryGroup",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of a registered query",
                        "name": "queryName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query to be executed",
                        "name": "queryBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results of the executed query",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tx/game/{txName}": {
            "post": {
                "description": "Submits a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Submits a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of a registered message",
                        "name": "txName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction details \u0026 message to be submitted",
                        "name": "txBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sign.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction hash and tick",
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.PostTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tx/persona/create-persona": {
            "post": {
                "description": "Creates a persona",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a persona",
                "parameters": [
                    {
                        "description": "Transaction details \u0026 message to be submitted",
                        "name": "txBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sign.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction hash and tick",
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.PostTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tx/{group}/{name}": {
            "post": {
                "description": "Submits a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Submits a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message group",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of a registered message",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction details \u0026 message to be submitted",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sign.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction hash and tick",
                        "schema": {
                            "$ref": "#/definitions/cardinal_server_handler.PostTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/world": {
            "get": {
                "description": "Contains the registered components, messages, queries, and namespace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves details of the game world",
                "responses": {
                    "200": {
                        "description": "Details of the game world",
                        "schema": {
                            "$ref": "#/definitions/pkg_world_dev_world-engine_cardinal_types.WorldInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cardinal_server_handler.GetHealthResponse": {
            "type": "object",
            "properties": {
                "isGameLoopRunning": {
                    "type": "boolean"
                },
                "isServerRunning": {
                    "type": "boolean"
                }
            }
        },
        "cardinal_server_handler.GetReceiptsRequest": {
            "type": "object",
            "properties": {
                "txHashes": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "cardinal_server_handler.GetReceiptsResponse": {
            "type": "object",
            "properties": {
                "receipts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "cardinal_server_handler.PostTransactionResponse": {
            "type": "object",
            "properties": {
                "txHash": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "pkg_world_dev_world-engine_cardinal_types.ComponentInfo": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "property name and type",
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "description": "name of the component",
                    "type": "string"
                }
            }
        },
        "pkg_world_dev_world-engine_cardinal_types.EndpointInfo": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "property name and type",
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "description": "name of the message or query",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "pkg_world_dev_world-engine_cardinal_types.EntityData": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "pkg_world_dev_world-engine_cardinal_types.WorldInfo": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_world_dev_world-engine_cardinal_types.ComponentInfo"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_world_dev_world-engine_cardinal_types.EndpointInfo"
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "queries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_world_dev_world-engine_cardinal_types.EndpointInfo"
                    }
                }
            }
        },
        "sign.Transaction": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "json string",
                    "type": "object"
                },
                "hash": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "personaTag": {
                    "type": "string"
                },
                "signature": {
                    "description": "hex encoded string",
                    "type": "string"
                }
            }
        }
    }
}